HSEARCH-4701 Move pojo-standalone Log class to package *.logging.impl
Because that's where it should be in order to work properly with our
module-info definition.
