+++ mapper/pojo-standalone/src/main/java/org/hibernate/search/mapper/pojo/standalone/logging/impl/StandalonePojoMappingHints.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.mapper.pojo.standalone.log.impl;
+package org.hibernate.search.mapper.pojo.standalone.logging.impl;
 
 import org.hibernate.search.engine.backend.mapping.spi.BackendMappingHints;
 import org.hibernate.search.util.common.logging.impl.MessageConstants;
